seperated routes, fixed merge bugs and removed some old code
