+++ package.json
@@ -18,6 +18,7 @@
     "express-session": "^1.16.1",
     "mongojs": "^2.6.0",
     "mongoose": "^5.5.12",
+    "multer": "^1.4.1",
     "slug": "^1.1.0"
   },
   "devDependencies": {
